# ChoboWeb/openapi/chobo.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Chobo API",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "tags": [
          "Logs"
        ],
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "last",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationLogEntryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationLogEntryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationLogEntryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logs/clear": {
      "post": {
        "tags": [
          "Logs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearApplicationLogsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearApplicationLogsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ClearApplicationLogsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/policies": {
      "get": {
        "tags": [
          "Policies"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupPolicyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupPolicyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupPolicyDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/inventory": {
      "get": {
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "sourceClusterId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyInventory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyInventory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyInventory"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/simulate": {
      "post": {
        "tags": [
          "Policies"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySimulationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySimulationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySimulationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySimulationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySimulationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySimulationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/policies/{id}": {
      "put": {
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupPolicyDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/policies/{id}/evaluate": {
      "post": {
        "tags": [
          "Policies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyEvaluationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyEvaluationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyEvaluationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyEvaluationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyEvaluationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyEvaluationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationLogEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exception": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BackupPolicyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "sourceClusterId": {
            "type": "string",
            "format": "uuid"
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "selectorJsonVersion": {
            "type": "integer",
            "format": "int32"
          },
          "selector": {
            "$ref": "#/components/schemas/PolicySelector"
          },
          "retention": {
            "$ref": "#/components/schemas/BackupRetentionDto"
          },
          "failedBackupRetentionMode": {
            "$ref": "#/components/schemas/FailedBackupRetentionMode"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BackupRetentionDto": {
        "type": "object",
        "properties": {
          "fullRetentionMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "incrementalRetentionMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "minBackupsToKeep": {
            "type": "integer",
            "format": "int32"
          },
          "minFullBackupsToKeep": {
            "type": "integer",
            "format": "int32"
          },
          "retentionMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ClearApplicationLogsRequest": {
        "type": "object",
        "properties": {
          "before": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "FailedBackupRetentionMode": {
        "enum": [
          "KeepAndExcludeFromMinBackupsToKeep",
          "DeleteByGarbageCollectorAfterFailure"
        ],
        "type": "string"
      },
...
      "PolicyEvaluationDto": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "format": "uuid"
          },
          "policyName": {
            "type": "string",
            "nullable": true
          },
          "sourceClusterId": {
            "type": "string",
            "format": "uuid"
          },
          "selectorJsonVersion": {
            "type": "integer",
            "format": "int32"
          },
          "selector": {
            "$ref": "#/components/schemas/PolicySelector"
          },
          "tables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyInventoryTable"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicyEvaluationRequest": {
        "type": "object",
        "properties": {
          "inventory": {
            "$ref": "#/components/schemas/PolicyInventory"
          }
        },
        "additionalProperties": false
      },
      "PolicyInventory": {
        "type": "object",
        "properties": {
          "tables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyInventoryTable"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicyInventoryTable": {
        "type": "object",
        "properties": {
          "database": {
            "type": "string",
            "nullable": true
          },
          "table": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicyMatchKind": {
        "enum": [
          "All",
          "Exact",
          "Wildcard"
        ],
        "type": "string"
      },
      "PolicySelector": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicySelectorRule"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicySelectorAction": {
        "enum": [
          "Include",
          "Exclude"
        ],
        "type": "string"
      },
      "PolicySelectorRule": {
        "type": "object",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PolicySelectorAction"
          },
          "database": {
            "$ref": "#/components/schemas/SelectorPattern"
          },
          "table": {
            "$ref": "#/components/schemas/SelectorPattern"
          }
        },
        "additionalProperties": false
      },
      "PolicySimulationDto": {
        "type": "object",
        "properties": {
          "sourceClusterId": {
            "type": "string",
            "format": "uuid"
          },
          "selector": {
            "$ref": "#/components/schemas/PolicySelector"
          },
          "inventory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyInventoryTable"
            },
            "nullable": true
          },
          "tables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyInventoryTable"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PolicySimulationRequest": {
        "type": "object",
        "properties": {
          "sourceClusterId": {
            "type": "string",
            "format": "uuid"
          },
          "selector": {
            "$ref": "#/components/schemas/PolicySelector"
          }
        },
        "additionalProperties": false
      },
...
      "SelectorPattern": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/PolicyMatchKind"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpsertPolicyRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "sourceClusterId": {
            "type": "string",
            "format": "uuid"
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "selector": {
            "$ref": "#/components/schemas/PolicySelector"
          },
          "retention": {
            "$ref": "#/components/schemas/BackupRetentionDto"
          },
          "failedBackupRetentionMode": {
            "$ref": "#/components/schemas/FailedBackupRetentionMode"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste a Chobo access token. Swagger UI sends it as a Bearer token.",
        "scheme": "bearer",
        "bearerFormat": "Opaque"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}